feat: implement user extraction middleware and secure blog routes with authentication

+++ server/controllers/blog.js
@@ -1,36 +1,27 @@
 const blogRouter = require('express').Router()
-const jwt = require('jsonwebtoken')
-const { request } = require('../app')
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const middleware = require('../utils/middleware')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+
+// const getTokenFrom = request => {
+//   const authorization = request.get('authorization')
+//   if (authorization && authorization.startsWith('Bearer ')) {
+//     return authorization.replace('Bearer ', '')
+//   }
+//   return null
+// }
 
 blogRouter.get('/', async (request, response) => {
   const blog = await Blog
-    .find({}).populate('user', {username: 1, name: 1, _id: 0})
+    .find({}).populate('user', {username: 1, name: 1})
 
     response.json(blog)
 })
 
-blogRouter.post('/', async (request, response) => {
+blogRouter.post('/', middleware.userExtractor, async (request, response) => {
+
+  const user = request.user
   const body = request.body
-
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
-
-  if (!user) {
-    return response.status(400).json({ error: 'userId missing or not valid' })
-  }
 
   const blog = new Blog({
     title: body.title,
@@ -60,12 +51,36 @@ blogRouter.get('/:id', async (request, response, next) => {
   }
 })
 
-blogRouter.delete('/:id', async (request, response) =>{
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  try {
+    const user = request.user
+
+    if (!user) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+      return response.status(404).json({ error: 'Blog not found' })
+    }
+
+    if (blog.user.toString() !== user.id.toString()) {
+      return response.status(401).json({ error: 'not authorized to delete this blog' })
+    }
+
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } catch (error) {
+    response.status(400).json({ error: 'invalid id or token' })
+  }
 })
 
-blogRouter.put('/:id', async (request, response) => {
+
+blogRouter.put('/:id', middleware.userExtractor, async (request, response) => {
+
+  const user = request.user
+  const body = request.body
+
   try {
     const updatedBlog = await Blog.findByIdAndUpdate(
       request.params.id,
@@ -77,7 +92,7 @@ blogRouter.put('/:id', async (request, response) => {
     } 
     response.status(200).json(updatedBlog);
   } catch (error) {
-    response.status(500).json({ message: error.message });
+    response.status(400).json({ error: 'invalid id or token' });
   }
 })
